requirements.txt
# Web Scraping & File Retrieval Agent Dependencies
# Phase 1 (MVP) Requirements

# Core web scraping
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
aiohttp==3.9.1

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.1

# Configuration & Data Validation
pydantic==2.5.0
pyyaml==6.0.1

# Async support
asyncio-throttle==1.0.2
aiofiles==23.2.1

# Logging & Monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Future LLM integration (Phase 2)
# openai==1.3.0
# langchain==0.0.340
